Move the jakarta.servlet related classes to a new package org.python.servlet
Adjust the inline documentation accordingly.
